@@ -1,55 +1,44 @@
 name: lectura
-description: A new Flutter project.
-
-publish_to: 'none' # Remove this line if you intend to publish to pub.dev
-
-version: 1.0.0+1
+description: A next-gen app for live transcription and smart exports.
+publish_to: 'none'
+version: 1.0.5+5
 
 environment:
-  sdk: ">=2.19.0 <4.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  speech_to_text: ^7.0.0 # Keep the existing version for stability
-  http: ^1.2.2 # Required for PayPal API integration
+  cupertino_icons: ^1.0.6
+
+  # Core Functionality
+  speech_to_text: ^6.6.2
+  http: ^1.2.2
+  in_app_purchase: ^3.0.9
+  url_launcher: ^6.3.1
+
+  # File Handling
   path_provider: ^2.0.15
-  open_filex: ^4.6.0 # Update to resolvable version
-  async: ^2.11.0
-  characters: ^1.3.0
-  clock: ^1.1.1
-  collection: ^1.18.0
-  http_parser: ^4.0.2 # Downgraded for compatibility
-  material_color_utilities: ^0.11.1
-  meta: ^1.15.0
-  path: ^1.9.0
-  string_scanner: ^1.2.0
-  app_links: ^6.3.3 # Stay with this version unless breaking changes are resolved
-  url_launcher: ^6.1.11
-  in_app_purchase: ^3.0.6 # Google Play Billing integration
-  flutter_dotenv: ^5.0.2 # Added for managing environment variables securely
-  rename: ^3.0.2
+  open_filex: ^4.7.0
+  pdf: ^3.10.8
+  docx_template: ^0.4.0
+
+  # Utilities
+  flutter_dotenv: ^5.1.0
+  logger: ^2.6.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  boolean_selector: ^2.1.1
-  fake_async: ^1.3.1
-  leak_tracker: ^10.0.5
-  leak_tracker_flutter_testing: ^3.0.5
-  matcher: ^0.12.16+1
-  stack_trace: ^1.11.1
-  test_api: ^0.7.2
-  vm_service: ^14.2.5 # Keep downgraded version for compatibility
-  flutter_launcher_icons: ^0.13.1
+  flutter_lints: ^3.0.2
+  mockito: ^5.4.4
 
-flutter_icons:
-  android: true
-  ios: false
-  image_path: "assets/icon/lectura_app_logo_512x512_transparent.png"
 
 flutter:
   uses-material-design: true
   assets:
-    - .env
+    - assets/icon/lectura_app_logo_512x512_transparent.png
+    - assets/empty.docx
+    - assets/template.docx
+    - assets/config/app.env
 
